requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
colorama==0.4.6
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.113.0
greenlet==3.0.3
h11==0.14.0
idna==3.8
pyasn1==0.6.0
pydantic==2.9.0
pydantic_core==2.23.2
PyMySQL==1.1.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.34
SQLAlchemy-Utils==0.41.2
starlette==0.38.4
typing_extensions==4.12.2
tzdata==2024.1
uvicorn==0.30.6
